+++ src/app/trip-info/search/trip-info-search.component.ts
@@ -6,7 +6,7 @@ import { SbbNotificationToast } from '@sbb-esta/angular/notification-toast';
 
 import * as OJP from 'ojp-sdk'
 
-import { APP_STAGE, DEBUG_LEVEL } from '../../config/constants'
+import { APP_STAGE, APP_STAGEs } from '../../config/constants'
 
 import { UserTripService } from 'src/app/shared/services/user-trip.service';
 import { TripInfoService } from '../trip-info.service';
@@ -53,13 +53,9 @@ export class TripInfoSearchComponent implements OnInit {
     this.model = {
       journeyRef: '',
       journeyDateTime: new Date(),
-      appStageOptions: ['PROD', 'INT', 'TEST', 'LA Beta'],
+      appStageOptions: APP_STAGEs,
       isSearching: false,
       permalinkURLAddress: '',
-    }
-
-    if (DEBUG_LEVEL === 'DEBUG') {
-      this.model.appStageOptions.push('OJP-SI');
     }
 
     this.updatePermalinkURLAddress();